requirements.txt
appdirs==1.4.3
click==6.7
fudge==1.1.0
future==0.16.0
packaging==16.8
pprint==0.1
pymongo==3.4.0
pyparsing==2.2.0
pyramda==0.1
python-dotenv==0.6.4
requests==2.13.0
simplejson==3.10.0
six==1.10.0
soundcloud==0.5.0
